# src/ApartmentManagement.API/V1/DTOs/Auth/RegisterRequestDto.cs
namespace ApartmentManagement.API.V1.DTOs.Auth;

public sealed class RegisterRequestDto
{
    public string Email { get; set; } = string.Empty;
    public string Password { get; set; } = string.Empty;
    public string FullName { get; set; } = string.Empty;
    public string? PhoneNumber { get; set; }
}