# swagger-collection.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apartment Management API v1.0",
    "version": "1.0.0",
    "description": "OpenAPI 3 — import vào Swagger Editor / UI. Mọi JSON body dùng schema có `example` trên từng field. Các UUID trong ví dụ có dạng ngẫu nhiên (giống `Guid.NewGuid()` khi seed) và **không** phải id thật trong DB — sau khi seed, thay bằng giá trị từ API. JWT: login rồi gửi Authorization: Bearer. Seed data range: apartments BLK-0001..BLK-0150, bulk users demoblk00..demoblk39@seed.local, canonical accounts admin1/admin2 and user1..user6 @ apartment.local. UUID examples are placeholders; after seeding, lấy ID thật từ API list/detail."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
    "/api/v1.0/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": []
      }
    },
    "/api/v1.0/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register (self-service)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation or identity error"
          }
        },
        "security": []
      }
    },
    "/api/v1.0/Auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": []
      }
    },
    "/api/v1.0/Auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": []
      }
    },
    "/api/v1.0/Auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": []
      }
    },
    "/api/v1.0/Auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current user (auth)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequestDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin3@apartment.local"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "ExampleFormat!1Abc"
          }
        }
      },
      "RegisterRequestDto": {
        "type": "object",
        "required": [
          "email",
          "password",
          "fullName"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "newuser@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Register!1Abc"
          },
          "fullName": {
            "type": "string",
            "example": "New Resident"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "example": "+84901234567"
          }
        }
      },
...
      "RefreshTokenRequestDto": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "base64-or-hex-refresh-token-string-from-login-response"
          }
        }
      },
      "ForgotPasswordRequestDto": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user6@apartment.local"
          }
        }
      },
      "ResetPasswordRequestDto": {
        "type": "object",
        "required": [
          "email",
          "token",
          "newPassword"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "user6@apartment.local"
          },
          "token": {
            "type": "string",
            "example": "reset-token-from-forgot-password-dev-response"
          },
          "newPassword": {
            "type": "string",
            "example": "NewStr0ng!Pass"
          }
        }
      },
...
    }
  }
}